Move all web-related stuff to a new module
