+++ web/http_auth.go
@@ -1,4 +1,4 @@
-package main
+package web
 
 import (
 	"github.com/bouncepaw/mycorrhiza/cfg"
@@ -20,14 +20,14 @@ func init() {
 }
 
 func handlerRegister(w http.ResponseWriter, rq *http.Request) {
-	prepareRq(rq)
+	util.PrepareRq(rq)
 	if !cfg.UseRegistration {
 		w.WriteHeader(http.StatusForbidden)
 	}
 	if rq.Method == http.MethodGet {
 		io.WriteString(
 			w,
-			base(
+			views.BaseHTML(
 				"Register",
 				views.RegisterHTML(rq),
 				user.FromRequest(rq),
@@ -61,7 +61,7 @@ func handlerLogout(w http.ResponseWriter, rq *http.Request) {
 		log.Println("Unknown user tries to log out")
 		w.WriteHeader(http.StatusForbidden)
 	}
-	w.Write([]byte(base("Logout?", views.LogoutHTML(can), u)))
+	w.Write([]byte(views.BaseHTML("Logout?", views.LogoutHTML(can), u)))
 }
 
 func handlerLogoutConfirm(w http.ResponseWriter, rq *http.Request) {
@@ -70,26 +70,26 @@ func handlerLogoutConfirm(w http.ResponseWriter, rq *http.Request) {
 }
 
 func handlerLoginData(w http.ResponseWriter, rq *http.Request) {
-	prepareRq(rq)
+	util.PrepareRq(rq)
 	var (
 		username = util.CanonicalName(rq.PostFormValue("username"))
 		password = rq.PostFormValue("password")
 		err      = user.LoginDataHTTP(w, rq, username, password)
 	)
 	if err != "" {
-		w.Write([]byte(base(err, views.LoginErrorHTML(err), user.EmptyUser())))
+		w.Write([]byte(views.BaseHTML(err, views.LoginErrorHTML(err), user.EmptyUser())))
 	} else {
 		http.Redirect(w, rq, "/", http.StatusSeeOther)
 	}
 }
 
 func handlerLogin(w http.ResponseWriter, rq *http.Request) {
-	prepareRq(rq)
+	util.PrepareRq(rq)
 	w.Header().Set("Content-Type", "text/html;charset=utf-8")
 	if user.AuthUsed {
 		w.WriteHeader(http.StatusOK)
 	} else {
 		w.WriteHeader(http.StatusForbidden)
 	}
-	w.Write([]byte(base("Login", views.LoginHTML(), user.EmptyUser())))
+	w.Write([]byte(views.BaseHTML("Login", views.LoginHTML(), user.EmptyUser())))
 }
